Messages centralized in mongo
The purpose of this is to no longer have to depend on the properties
files for handling customize error messaging for error codes like most
validation errors. This allows applications to scale out more easily
and to not have to worry about shared mounts. Also makes it easier to
develop out capabilities like configuration management through the
different environments.

Other benefits are

- Caching of responses for improved performance.
- More efficient iterator usage in mongo collection traversal.
- Better error handling on failed mongo invocations.
- Generic servlet filter for handling CORS compliance.
- Validation helper for non-blank values.