Refactorings and simplification of Mongo and Errors
Created a generic Identifiable interface to allow for a generic way for
retrieving the id of a bean without caring about any other details.

MongoDataStore is now able to handle the creation of a mongo instance
on its own without the need for a factory class. Allows for simplified
usage of the system with fewer moving/dependent parts.

BaseController now has a simplified model for rendering output. It uses
on a code level a direct output to the writer and a generic data bean
handler. This should lay a nice foundation for doing reusable
controllers and not having a score of jsp files. In fact the goal is to
remove the need for jsp entirely in the rendering of json or xml output