Monitor Component Added
This monitor can be used to quickly check if the application is running
and whether it is correctly speaking to the mongo server. Also returns
information about which server instance is the master and the name of
the replica set if applicable along with the version of mongo being hit.