Mapping annotations to reduce conversion code.
- Removed the 'replace' functionality for now. Felt like it was going
down the wrong path. Wanted to take a step back and rethink it a bit.

- Renamed MongoField to MongoMapping to make it a bit clearer.

- Fixed issue where primitives were not being handled correctly. Also
added char support as well.

- Added ability for conversion to handle subdocuments as well.

- MongoMapping now takes a "property" field as well. This allows you to
use a complex object but take only a specific property of that complex
object is used as the value in place of that field. This functionality
is strictly 'experimental' and is likely to be undone or adjusted a bit
further.