Configuration Improvements
Allowed the ability to retrieve a configuration and it be allowed to be
null without tossing an exception.

Also made the use of Configuration easier by not requiring a collection
name and polling frequency. By default 'config' is the collection name
and there is no polling (popular request for production) unless it is
explicitly stated as a property.