Dynamic Datasource
A class to enable utilizing relational datasources without having to
configure them in a properties or declaring the beans in spring. This
for example allows you to have the configurations in a Configuration
object which could be backed by a database and to dynamically create
and reconfigure the datasource during runtime without having to restart
the application.