Enhanced Configurations
Able to now save complex object configurations, maps, lists, and a
large variety of things as the value of a key now.

Enhanced the configuration rest controller to be a reduced number of
unique rest mappings making it a bit simplier to know and understand.

Updated Dynamic Data Source to have a Data Source Config which takes
advantage of the new enhanced configurations to try and standardize
the db configurations and allow atomic configuration modifications.

Added automated regression testing for configuration as it is not the
easiest thing to spot when an issue happens.