Mongo Configuration now honors type definition
Before if no explicit type was requested it would pull it in as a raw
map object. Now if it detects a _class in the object definition it will
attempt to convert it into that object before trying to convert it
into the requested type. This resolves the issue of the initial loading
pulling the configs into cache incorrectly and thus serving up the
wrong information.