DynamicDataSource no longer reloaded constantly
The DataSourceConfig did not have an equals implemented so during the
delta check it appeared that the database was always reconfigured so it
was constantly reconstructing the datasource each polling.

Longer term solution is to compare the document value previously to the
new document value pre-translation process to determine changes. Or to
perhaps attach a "modification" date to the system to help make things
a bit more efficient perhaps as well.