Fixed issue with Async parameter type recongition
The checking was reversed so a child of a class was not assignable to
the parent and thus not matching the method signatures. Updated so that
the parent is assignable to child instead (reversed) which is a true
statement now.

Also fixed a potential for a null pointer exception.

Slight issue potentially with null being able to match any parameter
and thus will likely cause issues in the future. Need to add some type
safety option to allow it to be worked around even if less cleanly.