Configuration Reload Hardening
There is a potential for a stale/closed mongo connection to cause the
configuration reloader to fail and prevent any future configurations
update checking until the system is restarted again.

A try/catch is used to log the incidents in a more noticable manner and
to trap the exception preventing the escalation from killing off the
poller which checks for updated configurations.