Configuration no longer exceptions when no key
This prevents the need to do a common implementation decision by almost
everyone to do .get(key, null, Class.class) so that if a value is not
found then a null is returned and something acts upon the null just as
if it were the expected value. The try/catch path creates for ugly code
that is difficult to follow.

It breaks backwards comaptiblity but that is why this codebase is still
in the 0.x phase so little things such as improved usability can be
added in without too much stress.