Enabled Callback Support for Asynchronous Calls
Now instead of manually having to "block" yourself to know when a
process invoked asynchronously has run, you can associate a callback
handler on the asynchronous calls. This allows you to know exactly
which methods threw an exception and/or was completed as the callback
is invoked appropriately for said task.