Added LoggingRestTemplate capability
A frequent need of many SOA systems is to "log" the actual request and
response results of any web service request. Along with that is the
status code and duration of the request for analytical reasons. Due to
the near constant request of information for web service requests, an
extension of the existing RestTemplate by spring was created to enable
a default logging capability for consistency and easy of use for all
systems.

Also enhanced the AjaxController with a standard error handling for a
non-authorized access to a service.