Only inject "errors" on an error response.
The dynamic injection of errors in all cases makes jQuery development
easier but it greatly increases development efforts of SOA API's. This
is because libraries claim to returne a well defined model but end up
having an extra field within it so when a recipient of a library uses
the same model to process the response it ends up breaking because it
no longer understands how to handle the "error" field.

Also angularJS and presumably newer javascript frameworks have a much
cleaner separation of errors and non error responses negating the need
to have all responses with a clean "errors" array to work with. If
jQuery users absolutely need to check all responses for errors, they
can check the existance of instead.