Do not warn on interruption of no work
Forceful closing of an asynchronous job when no work existed for it to
do became very noisy for those who wanted temporary asynchronous pools
for various reasons. The warning is important but not when there is no
harm to be had.

Also expanded it to make future debugging easier by allowing custom
thread namings.