Created Collection that may convert the data
This is especially useful when wrapping an existing collection without
having to make a brand new collection and storing all the results into
it. Thus things like responses from a DB are much faster and more
efficient because when they are streamed to an output stream in a data
transformed manner there is no need to maintain the intermediate step
anymore.